+++ lib/ZegoRenderView.d.ts
@@ -4,9 +4,11 @@ export interface SurfaceViewProps extends ViewProps {
     zOrderMediaOverlay?: boolean;
     zOrderOnTop?: boolean;
 }
+export interface TextureViewProps extends ViewProps {
+}
 export declare class ZegoSurfaceView extends Component<SurfaceViewProps, {}> {
     render(): any;
 }
-export declare class ZegoTextureView extends Component<{}> {
+export declare class ZegoTextureView extends Component<TextureViewProps, {}> {
     render(): any;
 }